Move implementation to an internal header to allow testing
