+++ test/gtest/CMakeLists.txt
@@ -1,7 +1,9 @@
 include_directories(../../third_party/googletest/googletest/include)
+include_directories(../../src/wasm)
 
 set(unittest_SOURCES
   type-builder.cpp
+  wat-parser.cpp
 )
 
 binaryen_add_executable(binaryen-unittests "${unittest_SOURCES}")
